Add a flying pathfinding solution to make navigation better  (#39)

+++ Assets/Digital Painting/Scripts/Editor/PackageBuilder.cs
@@ -5,7 +5,9 @@ using UnityEditor;
 using UnityEngine;
 
 /// <summary>
-/// Build the package for release.
+/// Build the package for release. Grabs all the available plugin packages
+/// and puts them in the `plugins` folder then builds the Unity Package
+/// for The Digital Painting.
 /// 
 /// Run from Unity with `Digital Painting -> Build -> Build Core Package`
 /// 
@@ -15,40 +17,52 @@ using UnityEngine;
 public class PackageBuilder
 {
 
-    [MenuItem("Digital Painting/Build/Build Core Package")]
+    [MenuItem("Tools/Build Package/Digital Painting/Core")]
     public static void Build()
     {
-        string rootDir = "Assets\\Digital Painting";
+        string[] rootDirs = { "Assets\\Digital Painting", "Assets\\Flying-Pathfinding", "Assets\\Gizmos\\DigitalPainting" };
         string excludeSubDir = "Plugins";
-        string packageName = "DigitalPainting.unitypackage";
+        string packageName = @"..\DigitalPainting.unitypackage";
+
+        // Collect available plugins
+        GetPlugins();
 
         // Delete everything in plugins directory except *.unitypackage and *.md (and matching .meta)
-        MoveExcludedFiles(rootDir + "\\" + excludeSubDir);
+        foreach (string rootDir in rootDirs) {
+            MoveExcludedFiles(rootDir + "\\" + excludeSubDir);
+        }
         AssetDatabase.Refresh();
 
-        AssetDatabase.ExportPackage(rootDir, packageName, ExportPackageOptions.Interactive | ExportPackageOptions.Recurse);
+        AssetDatabase.ExportPackage(rootDirs, packageName, ExportPackageOptions.Interactive | ExportPackageOptions.Recurse);
         Debug.Log("Exported " + packageName);
 
-        RecoverExcludedFiles(rootDir + "\\" + excludeSubDir);
+        foreach (string rootDir in rootDirs)
+        {
+            RecoverExcludedFiles(rootDir + "\\" + excludeSubDir);
+        }
         AssetDatabase.Refresh();
     }
 
     protected static void MoveExcludedFiles(string dir)
     {
-        string[] subdirectoryEntries = Directory.GetDirectories(dir);
-        foreach (string subdirectory in subdirectoryEntries)
+        if (File.Exists(dir))
         {
-            if (Path.GetFileName(subdirectory) != "Scenes")
+            string[] subdirectoryEntries = Directory.GetDirectories(dir);
+            foreach (string subdirectory in subdirectoryEntries)
             {
-                Debug.Log("Moving to safety: " + subdirectory);
-                string copyPath = "Temp" + Path.DirectorySeparatorChar + subdirectory;
-                Directory.CreateDirectory(Path.GetDirectoryName(copyPath));
-                Directory.Move(subdirectory, copyPath);
-                File.Move(subdirectory + ".meta", copyPath + ".meta");
-            }
-            else
-            {
-                MoveExcludedFiles(subdirectory);
+                if (Path.GetFileName(subdirectory) != "Scenes")
+                {
+                    Debug.Log("Moving to safety: " + subdirectory);
+                    string copyPath = "Temp" + Path.DirectorySeparatorChar + subdirectory;
+                    Directory.CreateDirectory(Path.GetDirectoryName(copyPath));
+                    Directory.Move(subdirectory, copyPath);
+                    File.Move(subdirectory + ".meta", copyPath + ".meta");
+                    Debug.Log("Temporarily moved " + subdirectory);
+                }
+                else
+                {
+                    MoveExcludedFiles(subdirectory);
+                }
             }
         }
     }
@@ -56,23 +70,44 @@ public class PackageBuilder
     protected static void RecoverExcludedFiles(string dir)
     {
         string copyPath = "Temp" + Path.DirectorySeparatorChar + dir;
-        string[] subdirectoryEntries = Directory.GetDirectories(copyPath);
-        foreach (string subdirectory in subdirectoryEntries)
+        if (File.Exists(copyPath))
         {
-            if (Path.GetFileName(subdirectory) != "Scenes")
+            string[] subdirectoryEntries = Directory.GetDirectories(copyPath);
+            foreach (string subdirectory in subdirectoryEntries)
             {
                 string targetPath = subdirectory.Substring(subdirectory.IndexOf(Path.DirectorySeparatorChar, 1) + 1);
-                Debug.Log("Moving back to project from: " + subdirectory + " to " + targetPath);
-                Directory.CreateDirectory(Path.GetDirectoryName(targetPath));
-                Directory.Move(subdirectory, targetPath);
-                File.Move(subdirectory + ".meta", targetPath + ".meta");
+                if (Path.GetFileName(subdirectory) != "Scenes")
+                {
+                    Debug.Log("Moving back to project from: " + subdirectory + " to " + targetPath);
+                    Directory.CreateDirectory(Path.GetDirectoryName(targetPath));
+                    Directory.Move(subdirectory, targetPath);
+                    File.Move(subdirectory + ".meta", targetPath + ".meta");
+                    Debug.Log("Moved back " + targetPath);
+                }
+                else
+                {
+                    RecoverExcludedFiles(targetPath);
+                }
             }
-            else
-            {
-                MoveExcludedFiles(subdirectory);
+            Directory.Delete(copyPath, true);
+        }
+    }
+
+    protected static void GetPlugins()
+    {
+        string[] files =  Directory.GetFiles("..", "*.unitypackage");
+        foreach (string file in files)
+        {
+            if (file != "..\\DigitalPainting.unitypackage") {
+                string targetPath = "Assets\\Digital Painting\\Plugins\\" + file.Substring(file.IndexOf(Path.DirectorySeparatorChar, 1) + 1);
+                if (File.Exists(targetPath))
+                {
+                    File.Delete(targetPath);
+                }
+                File.Move(file, targetPath);
+                Debug.Log("Imported plugin package " + file + " to " + targetPath);
             }
         }
-        Directory.Delete(copyPath, true);
     }
 
 }
